pyproject.toml
[project]
name = "cert-graveyard-yara"
version = "1.0.0"
description = "Automated YARA rule generation from CertGraveyard compromised certificate database"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CertGraveyard YARA Generator"}
]
dependencies = [
    "httpx>=0.27.0",
    "pandas>=2.0.0",
    "jinja2>=3.1.0",
    "typer>=0.12.0",
    "yara-python>=4.5.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pandas-stubs>=2.0.0",
    "respx>=0.21.0",
]

[project.scripts]
cert-graveyard-yara = "cert_graveyard_yara.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cert_graveyard_yara"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--cov=src/cert_graveyard_yara",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=65"
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "yara"
ignore_missing_imports = true

